requirements.test.txt
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version

aioresponses
algoliasearch==2.2.0
codecov==2.1.12
coverage==6.2.0
mypy==0.931
pytest-cov==2.12.1
pytest-homeassistant-custom-component>=0.6.2
